Added error checking
Added error checking to the add,edit and delete stock GUI's as well as the delete sale GUI and fixed any problematic naming

+++ PHP/PHP/DeleteItem.cs
@@ -32,40 +32,42 @@ namespace PHP
             {
                 string[] row = { p.ProductId.ToString(), p.Product_Name.ToString(), p.Price.ToString(), p.Stock_Level.ToString() };
                 var listViewItem = new ListViewItem(row);
-                listView1.Items.Add(listViewItem);
+                StockList.Items.Add(listViewItem);
             };
         }
-        private void button2_Click(object sender, EventArgs e)
+        private void RemoveButton_Click(object sender, EventArgs e)
         {
             int ID;
-            int.TryParse(textBox1.Text, out ID);
-            Product _Product = _PHPRepo.GetProductbyId(ID);
-            _PHPRepo.deleteProduct(_Product);
-            MessageBox.Show("Product has been deleted ", "Sucessful", MessageBoxButtons.OK, MessageBoxIcon.None);
-            listView1.Items.Clear();
-            List<Product> _NewList = _PHPRepo.GetProducts();
-            foreach (Product p in _NewList)
+            if (int.TryParse(ProductID.Text, out ID))
             {
-                string[] row = { p.ProductId.ToString(), p.Product_Name.ToString(), p.Price.ToString(), p.Stock_Level.ToString() };
-                var listViewItem = new ListViewItem(row);
-                listView1.Items.Add(listViewItem);
-            };
+                if (_PHPRepo.VarifyProductID(ID))
+                {
+                    _PHPRepo.GetProductbyId(ID);
+                    Product _Product = _PHPRepo.GetProductbyId(ID);
+                    _PHPRepo.deleteProduct(_Product);
+                    MessageBox.Show("Product has been deleted ", "Sucessful", MessageBoxButtons.OK, MessageBoxIcon.None);
+                    StockList.Items.Clear();
+                    List<Product> _NewList = _PHPRepo.GetProducts();
+                    foreach (Product p in _NewList)
+                    {
+                        string[] row = { p.ProductId.ToString(), p.Product_Name.ToString(), p.Price.ToString(), p.Stock_Level.ToString() };
+                        var listViewItem = new ListViewItem(row);
+                        StockList.Items.Add(listViewItem);
+                    };
+                    ProductID.Clear();
+                }
+                else
+                {
+                    MessageBox.Show("Could not find item: " + ID);
+                }
+            }
+            else
+            {
+                MessageBox.Show("Please enter a valid ID");
+            }
         }
 
-        private void label1_Click(object sender, EventArgs e)
-        {
 
-        }
-
-        private void listView1_SelectedIndexChanged(object sender, EventArgs e)
-        {
-
-        }
-
-        private void DeleteItem_Load(object sender, EventArgs e)
-        {
-
-        }
 
         private void CloseButton_Click(object sender, EventArgs e)
         {
@@ -81,6 +83,20 @@ namespace PHP
             {
                 //do nothing after dialog box is closed
             }
+        }
+        private void label1_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void listView1_SelectedIndexChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void DeleteItem_Load(object sender, EventArgs e)
+        {
+
         }
     }
 }
